+++ examples/default/__tests__/e2e/pageobjects/ExampleSearchBoxHitsPage.js
