+++ examples/default/src/examples/ExampleSearchBoxHits.css
@@ -1,4 +1,4 @@
-@import './static/fonts/fonts.css';
+@import '../static/fonts/fonts.css';
 
 .cvi-SearchBox-input {
   height: 48px;
