+++ examples/default/src/examples/ExampleSearchBoxHits.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import ExampleSearchBoxHits from './ExampleSearchBoxHits';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<ExampleSearchBoxHits />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
